H2 implemented, registertest fixed, keycloak disable for register added, custom kecloak realm setup for testing
