+++ src/main/resources/application.properties
@@ -8,10 +8,6 @@ keycloak.credentials.secret=Z8kzdqRzPcfWZENlvPebAo3UCjeiQ0UZ
 master.keycloak.username=temp
 master.keycloak.password=temp
 
-#spring.security.oauth2.client.provider.external.issuer-uri=https://keycloak.switchfully.com/realms/java-2023-10
-#spring.security.oauth2.client.registration.external.client-id=switchfully-lms
-#spring.security.oauth2.client.registration.external.client-secret=Z8kzdqRzPcfWZENlvPebAo3UCjeiQ0UZ
-
 spring.datasource.url=jdbc:postgresql://database-2.cuvv0osxzgmi.eu-west-3.rds.amazonaws.com:5432/lms_java_10_2023?currentSchema=lms_java_10_2023_schema
 spring.datasource.username=student
 spring.datasource.password=spidermanismyhero
@@ -19,4 +15,3 @@ spring.datasource.driverClassName=org.postgresql.Driver
 spring.jpa.hibernate.ddl-auto=validate
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
-
