+++ src/test/resources/application.properties
@@ -1,0 +1,19 @@
+spring.security.oauth2.resourceserver.jwt.issuer-uri=https://keycloak.switchfully.com/realms/switchfully-lms-test
+keycloak.auth-server-url=https://keycloak.switchfully.com
+keycloak.realm=switchfully-lms-test
+keycloak.resource=switchfully-lms-test-client
+keycloak.credentials.secret=rcb6nKkHvjxqqWsBIVrelsN6SysGEXD3
+
+master.keycloak.username=temp
+master.keycloak.password=temp
+
+spring.datasource.url=jdbc:h2:mem:test;MODE=PostgreSQL;
+spring.datasource.username=
+spring.datasource.password=
+
+server.error.include-message=always
+spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.defer-datasource-initialization=true
+
+spring.sql.init.mode=always
+spring.flyway.enabled=false
