+++ src/test/resources/data.sql
@@ -1,0 +1,21 @@
+INSERT INTO progresses (name)
+VALUES
+    ('DONE'),
+    ('REFACTORING'),
+    ('TESTING'),
+    ('STUCK'),
+    ('FEEDBACK_NEEDED'),
+    ('BUSY'),
+    ('NOT_STARTED');
+
+
+INSERT INTO users (dtype, email, display_name)
+values
+    ('Student', 'Balder@lms.com', 'Balder'),
+    ('Student', 'Dylan@lms.com', 'Dylan'),
+    ('Student', 'Karel@lms.com', 'Karel'),
+    ('Student', 'Jussi@lms.com', 'Jussi'),
+    ('Student', 'Thomas@lms.com', 'Thomas'),
+    ('Student', 'Manon@lms.com', 'Manon'),
+    ('Student', 'Sandro@lms.com', 'Sandro'),
+    ('Student', 'Wietse@lms.com', 'Wietse')
