+++ src/test/resources/schema.sql
@@ -1,0 +1,81 @@
+CREATE TABLE IF NOT EXISTS users (
+    id                  BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
+    dtype               VARCHAR(31),
+    email               VARCHAR(255),
+    display_name        VARCHAR(255),
+    CONSTRAINT pk_users PRIMARY KEY (id)
+);
+
+CREATE TABLE IF NOT EXISTS courses (
+    id                  BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
+    name                VARCHAR(255),
+    CONSTRAINT pk_courses PRIMARY KEY (id)
+);
+
+CREATE TABLE IF NOT EXISTS class_groups (
+    id                  BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
+    name                VARCHAR(255),
+    course_id           BIGINT,
+    start_date          DATE DEFAULT NULL,
+    end_date            DATE DEFAULT NULL,
+    CONSTRAINT pk_class_groups PRIMARY KEY (id),
+    CONSTRAINT FK_CLASS_GROUPS_ON_COURSE FOREIGN KEY (course_id) REFERENCES courses (id)
+);
+
+CREATE TABLE IF NOT EXISTS class_groups_coaches (
+    class_group_id      BIGINT NOT NULL,
+    coach_id            BIGINT NOT NULL,
+    CONSTRAINT fk_clagrocoa_on_class_group FOREIGN KEY (class_group_id) REFERENCES class_groups (id),
+    CONSTRAINT fk_clagrocoa_on_coach FOREIGN KEY (coach_id) REFERENCES users (id)
+);
+
+CREATE TABLE IF NOT EXISTS class_groups_students (
+    class_group_id      BIGINT NOT NULL,
+    student_id          BIGINT NOT NULL,
+    CONSTRAINT fk_clagrostu_on_class_group FOREIGN KEY (class_group_id) REFERENCES class_groups (id),
+    CONSTRAINT fk_clagrostu_on_student FOREIGN KEY (student_id) REFERENCES users (id)
+);
+
+CREATE TABLE IF NOT EXISTS modules (
+    id                  BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
+    dtype               VARCHAR(31),
+    name                VARCHAR(255),
+    parent_module_id    BIGINT,
+    CONSTRAINT pk_modules PRIMARY KEY (id),
+    CONSTRAINT FK_MODULES_ON_PARENT_MODULE FOREIGN KEY (parent_module_id) REFERENCES modules (id)
+);
+
+CREATE TABLE IF NOT EXISTS codelabs (
+    id                  BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
+    name                VARCHAR(255),
+    module_id           BIGINT,
+    CONSTRAINT pk_codelabs PRIMARY KEY (id),
+    CONSTRAINT FK_CODELABS_ON_MODULE FOREIGN KEY (module_id) REFERENCES modules (id)
+);
+
+CREATE TABLE IF NOT EXISTS progresses (
+    id                  BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
+    name                VARCHAR(255),
+    CONSTRAINT pk_progresses PRIMARY KEY (id)
+);
+
+CREATE TABLE IF NOT EXISTS codelab_progresses (
+    id                  BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
+    codelab_id          BIGINT,
+    progress_id         BIGINT,
+    student_id          BIGINT,
+    CONSTRAINT pk_codelab_progresses PRIMARY KEY (id),
+    CONSTRAINT FK_CODELAB_PROGRESSES_ON_CODELAB FOREIGN KEY (codelab_id) REFERENCES codelabs (id),
+    CONSTRAINT FK_CODELAB_PROGRESSES_ON_PROGRESS FOREIGN KEY (progress_id) REFERENCES progresses (id),
+    CONSTRAINT FK_CODELAB_PROGRESSES_ON_STUDENT FOREIGN KEY (student_id) REFERENCES users (id)
+);
+
+CREATE TABLE IF NOT EXISTS comments (
+    id                  BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
+    text                TEXT,
+    user_id             BIGINT,
+    codelab_id          BIGINT,
+    CONSTRAINT pk_comments PRIMARY KEY (id),
+    CONSTRAINT FK_COMMENTS_ON_CODELAB FOREIGN KEY (codelab_id) REFERENCES codelabs (id),
+    CONSTRAINT FK_COMMENTS_ON_USER FOREIGN KEY (user_id) REFERENCES users (id)
+);
